pyproject.toml
[tool.poetry]
name = "drink3"
version = "0.0.0"
description = "23.05.08. 내배캠 drf 프로젝트"
authors = ["Tae seokyoung <xotjrud@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.2.2"
pillow = "^9.5.0"
pylint = "^2.17.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
